+++ projects.go
@@ -3,14 +3,11 @@ package contempt
 import (
 	"errors"
 	"fmt"
-	"io"
 	"io/fs"
 	"os"
 	"path/filepath"
-	"reflect"
 	"sort"
 	"strings"
-	"text/template"
 )
 
 // FindProjects returns a slice of all images that can be built from this repo, sorted such that images are positioned
@@ -72,39 +69,20 @@ func FindProjects(dir, templateName string) ([]string, error) {
 }
 
 func dependencies(dir, templateName string) []string {
-	var res []string
-	fakeFunks := template.FuncMap{}
-	for i := range templateFuncs {
-		for f := range templateFuncs[i] {
-			out := reflect.ValueOf(templateFuncs[i][f]).Type().Out(0).Kind()
-			if f == "image" {
-				fakeFunks[f] = func(dep string) string {
-					// Ignore fully-qualified images like "docker.io/library/alpine"
-					if index := strings.IndexByte(dep, '.'); index == -1 || index > strings.IndexByte(dep, '/') {
-						res = append(res, dep)
-					}
-					return ""
-				}
-			} else if out == reflect.Map {
-				fakeFunks[f] = func(args ...string) map[string]string {
-					return nil
-				}
-			} else if out == reflect.Slice {
-				fakeFunks[f] = func(args ...string) []string {
-					return nil
-				}
-			} else {
-				fakeFunks[f] = func(args ...string) string {
-					return ""
-				}
-			}
-		}
+	templatePath := filepath.Join(dir, templateName)
+
+	calls, err := engine.DryRun(templatePath)
+	if err != nil {
+		panic(err)
 	}
 
-	templatePath := filepath.Join(dir, templateName)
-	tpl := template.New(templatePath)
-	tpl.Funcs(fakeFunks)
-	_, _ = tpl.ParseFiles(templatePath)
-	_ = tpl.ExecuteTemplate(io.Discard, templateName, nil)
+	var res []string
+	for i := range calls["image"] {
+		dep := calls["image"][i][0]
+		if index := strings.IndexByte(dep, '.'); index == -1 || index > strings.IndexByte(dep, '/') {
+			res = append(res, dep)
+		}
+	}
+	
 	return res
 }